+++ csharpsrc/Tests/LeetCodeCases/No23.MergeKSortedLists/Constraints.cs
@@ -1,0 +1,42 @@
+﻿// <copyright file="Constraints.cs" company="Ocean">
+// Copyright (c) Ocean. All rights reserved.
+// </copyright>
+
+namespace Tests.LeetCodeCases.No23.MergeKSortedLists
+{
+    /// <summary>
+    /// The constraints of LeetCode No.23 question: Merge K Sorted Lists.
+    /// </summary>
+    internal class Constraints
+    {
+        /// <summary>
+        /// The min count of list nodes.
+        /// </summary>
+        public const int MinListNodeCount = 0;
+
+        /// <summary>
+        /// The max count of list nodes.
+        /// </summary>
+        public const int MaxListNodeCount = 10000;
+
+        /// <summary>
+        /// The min length of list.
+        /// </summary>
+        public const int MinLength = 0;
+
+        /// <summary>
+        /// The max length of list.
+        /// </summary>
+        public const int MaxLength = 500;
+
+        /// <summary>
+        /// The min value.
+        /// </summary>
+        public const int MinValue = -10000;
+
+        /// <summary>
+        /// The max value.
+        /// </summary>
+        public const int MaxValue = 10000;
+    }
+}